fix: correct reps data type casting and exercise ID validation
- Cast reps to INTEGER in SQL function (was missing)
- Validate exerciseId is valid UUID before sending
- Parse reps to integers, handle string inputs
- Handle rep ranges by taking first number

+++ database/fix-workout-transaction-id-mapping.sql
@@ -218,7 +218,7 @@ BEGIN
         NULLIF(v_exercise->>'exercise_id', '')::UUID,
         v_exercise->>'exercise_name',
         (v_exercise->>'sets')::INTEGER,
-        v_exercise->>'reps',
+        (v_exercise->>'reps')::INTEGER,
         v_exercise->>'weight_type',
         (v_exercise->>'weight')::NUMERIC,
         (v_exercise->>'weight_max')::NUMERIC,
